+++ src/anemoi/graphs/commands/inspect.py
@@ -1,7 +1,16 @@
+# (C) Copyright 2024 Anemoi contributors.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
 import argparse
 
-from anemoi.graphs.descriptor import GraphDescriptor
-from anemoi.graphs.inspector import GraphInspector
+from anemoi.graphs.describe import GraphDescriptor
+from anemoi.graphs.inspect import GraphInspector
 
 from . import Command