+++ src/anemoi/graphs/nodes/builders/from_refined_icosahedron.py
@@ -1,3 +1,12 @@
+# (C) Copyright 2024 Anemoi contributors.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
 from __future__ import annotations
 
 import logging
@@ -7,11 +16,11 @@ from abc import abstractmethod
 import networkx as nx
 import numpy as np
 import torch
-from anemoi.utils.config import DotDict
 from torch_geometric.data import HeteroData
 
 from anemoi.graphs.generate.hex_icosahedron import create_hex_nodes
 from anemoi.graphs.generate.masks import KNNAreaMaskBuilder
+from anemoi.graphs.generate.tri_icosahedron import create_stretched_tri_nodes
 from anemoi.graphs.generate.tri_icosahedron import create_tri_nodes
 from anemoi.graphs.nodes.builders.base import BaseNodeBuilder
 
@@ -25,8 +34,6 @@ class IcosahedralNodes(BaseNodeBuilder, ABC):
     ----------
     resolution : list[int] | int
         Refinement level of the mesh.
-    name : str
-        Name of the nodes.
     """
 
     def __init__(
@@ -40,6 +47,12 @@ class IcosahedralNodes(BaseNodeBuilder, ABC):
             self.resolutions = resolution
 
         super().__init__(name)
+        self.hidden_attributes = BaseNodeBuilder.hidden_attributes | {
+            "resolutions",
+            "nx_graph",
+            "node_ordering",
+            "area_mask_builder",
+        }
 
     def get_coordinates(self) -> torch.Tensor:
         """Get the coordinates of the nodes.
@@ -55,20 +68,13 @@ class IcosahedralNodes(BaseNodeBuilder, ABC):
     @abstractmethod
     def create_nodes(self) -> tuple[nx.DiGraph, np.ndarray, list[int]]: ...
 
-    def register_attributes(self, graph: HeteroData, config: DotDict) -> HeteroData:
-        graph[self.name]["_resolutions"] = self.resolutions
-        graph[self.name]["_nx_graph"] = self.nx_graph
-        graph[self.name]["_node_ordering"] = self.node_ordering
-        graph[self.name]["_aoi_mask_builder"] = self.aoi_mask_builder
-        return super().register_attributes(graph, config)
-
 
 class LimitedAreaIcosahedralNodes(IcosahedralNodes):
     """Nodes based on iterative refinements of an icosahedron using an area of interest.
 
     Attributes
     ----------
-    aoi_mask_builder : KNNAreaMaskBuilder
+    area_mask_builder : KNNAreaMaskBuilder
         The area of interest mask builder.
     """
 
@@ -83,10 +89,10 @@ class LimitedAreaIcosahedralNodes(IcosahedralNodes):
 
         super().__init__(resolution, name)
 
-        self.aoi_mask_builder = KNNAreaMaskBuilder(reference_node_name, margin_radius_km, mask_attr_name)
+        self.area_mask_builder = KNNAreaMaskBuilder(reference_node_name, margin_radius_km, mask_attr_name)
 
     def register_nodes(self, graph: HeteroData) -> None:
-        self.aoi_mask_builder.fit(graph)
+        self.area_mask_builder.fit(graph)
         return super().register_nodes(graph)
 
 
@@ -117,12 +123,12 @@ class LimitedAreaTriNodes(LimitedAreaIcosahedralNodes):
 
     Parameters
     ----------
-    aoi_mask_builder: KNNAreaMaskBuilder
+    area_mask_builder: KNNAreaMaskBuilder
         The area of interest mask builder.
     """
 
     def create_nodes(self) -> tuple[nx.Graph, np.ndarray, list[int]]:
-        return create_tri_nodes(resolution=max(self.resolutions), aoi_mask_builder=self.aoi_mask_builder)
+        return create_tri_nodes(resolution=max(self.resolutions), area_mask_builder=self.area_mask_builder)
 
 
 class LimitedAreaHexNodes(LimitedAreaIcosahedralNodes):
@@ -132,9 +138,54 @@ class LimitedAreaHexNodes(LimitedAreaIcosahedralNodes):
 
     Parameters
     ----------
-    aoi_mask_builder: KNNAreaMaskBuilder
+    area_mask_builder: KNNAreaMaskBuilder
         The area of interest mask builder.
     """
 
     def create_nodes(self) -> tuple[nx.Graph, np.ndarray, list[int]]:
-        return create_hex_nodes(resolution=max(self.resolutions), aoi_mask_builder=self.aoi_mask_builder)
+        return create_hex_nodes(resolution=max(self.resolutions), area_mask_builder=self.area_mask_builder)
+
+
+class StretchedIcosahedronNodes(IcosahedralNodes):
+    """Nodes based on iterative refinements of an icosahedron with 2
+    different resolutions.
+
+    Attributes
+    ----------
+    area_mask_builder : KNNAreaMaskBuilder
+        The area of interest mask builder.
+    """
+
+    def __init__(
+        self,
+        global_resolution: int,
+        lam_resolution: int,
+        name: str,
+        reference_node_name: str,
+        mask_attr_name: str,
+        margin_radius_km: float = 100.0,
+    ) -> None:
+
+        super().__init__(lam_resolution, name)
+        self.global_resolution = global_resolution
+
+        self.area_mask_builder = KNNAreaMaskBuilder(reference_node_name, margin_radius_km, mask_attr_name)
+
+    def register_nodes(self, graph: HeteroData) -> None:
+        self.area_mask_builder.fit(graph)
+        return super().register_nodes(graph)
+
+
+class StretchedTriNodes(StretchedIcosahedronNodes):
+    """Nodes based on iterative refinements of an icosahedron with 2
+    different resolutions.
+
+    It depends on the trimesh Python library.
+    """
+
+    def create_nodes(self) -> tuple[nx.Graph, np.ndarray, list[int]]:
+        return create_stretched_tri_nodes(
+            base_resolution=self.global_resolution,
+            lam_resolution=max(self.resolutions),
+            area_mask_builder=self.area_mask_builder,
+        )