+++ src/anemoi/graphs/plotting/interactive_html.py
@@ -1,3 +1,12 @@
+# (C) Copyright 2024 Anemoi contributors.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
 import logging
 from pathlib import Path
 from typing import Optional
@@ -6,6 +15,7 @@ from typing import Union
 import matplotlib.pyplot as plt
 import numpy as np
 import plotly.graph_objects as go
+from matplotlib.colors import rgb2hex
 from torch_geometric.data import HeteroData
 
 from anemoi.graphs.plotting.prepare import compute_isolated_nodes
@@ -136,7 +146,7 @@ def plot_isolated_nodes(graph: HeteroData, out_file: Optional[Union[str, Path]] 
                 mode="markers",
                 hoverinfo="text",
                 name=name,
-                marker={"showscale": False, "color": colorbar[len(nodes)], "size": 10},
+                marker={"showscale": False, "color": rgb2hex(colorbar[len(nodes)]), "size": 10},
             ),
         )