+++ src/anemoi/graphs/processors/__init__.py
@@ -1,0 +1,3 @@
+from .post_process import RemoveUnconnectedNodes
+
+__all__ = [RemoveUnconnectedNodes]