+++ CHANGELOG.md
@@ -8,20 +8,61 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please add your functional changes to the appropriate section in the PR.
 Keep it human-readable, your future self will thank you!
 
-## [Unreleased](https://github.com/ecmwf/anemoi-graphs/compare/0.3.0...HEAD)
+## [Unreleased](https://github.com/ecmwf/anemoi-graphs/compare/0.4.1...HEAD)
+
+## [0.4.1 - ICON graphs, multiple edge builders and post processors](https://github.com/ecmwf/anemoi-graphs/compare/0.4.0...0.4.1) - 2024-11-26
+
+### Added
+
+- feat: Define node sets and edges based on an ICON icosahedral mesh (#53)
+- feat: Add support for `post_processors` in the recipe. (#71)
+- feat: Add `RemoveUnconnectedNodes` post processor to clean unconnected nodes in LAM. (#71)
+- feat: Define node sets and edges based on an ICON icosahedral mesh (#53)
+- feat: Support for multiple edge builders between two sets of nodes (#70)
+
+# Changed
+
+- fix: bug when computing area weights with scipy.Voronoi. (#79)
+
+## [0.4.0 - LAM and stretched graphs](https://github.com/ecmwf/anemoi-graphs/compare/0.3.0...0.4.0) - 2024-11-08
 
 ### Added
+
 - ci: hpc-config, CODEOWNERS (#49)
 - feat: New node builder class, CutOutZarrDatasetNodes, to create nodes from 2 datasets. (#30)
 - feat: New class, KNNAreaMaskBuilder, to specify Area of Interest (AOI) based on a set of nodes. (#30)
 - feat: New node builder classes, LimitedAreaXXXXXNodes, to create nodes within an Area of Interest (AOI). (#30)
 - feat: Expanded MultiScaleEdges to support multi-scale connections in limited area graphs. (#30)
 - feat: New edge attribute, Azimuth. (#54)
+- feat: New method update_graph(graph) in the GraphCreator class. (#60)
+- feat: New class StretchedTriNodes to create a stretched mesh. (#51)
+- feat: Expanded MultiScaleEdges to support multi-scale connections in stretched graphs. (#51)
+- fix: bug in color when plotting isolated nodes (#63)
+- Add anemoi-transform link to documentation (#59)
+- Added `CutOutMask` class to create a mask for a cutout. (#68)
+- Added `MissingZarrVariable` and `NotMissingZarrVariable` classes to create a mask for missing zarr variables. (#68)
+- feat: Add CONTRIBUTORS.md file. (#72)
+- Create package documentation.
 
 ### Changed
+
 - ci: small fixes and updates pre-commit, downsteam-ci (#49)
 - feat: Implementing KNNEdges and CutOffEdges with torch-cluster (#54)
 - feat: Migrate edge attributes to inherit from MessagePassing (#54)
+- Update CODEOWNERS (#61)
+- ci: extened python versions to include 3.11 and 3.12 (#66)
+- Update copyright notice (#67)
+
+### Removed
+
+- Remove `CutOutZarrDatasetNodes` class. (#68)
+- Update CODEOWNERS
+- Fix pre-commit regex
+- ci: extened python versions to include 3.11 and 3.12
+- Update copyright notice
+- Fix `__version__` import in init
+- The `edge_builder` field in the recipe is renamed to `edge_builders`. It now receives a list of edge builders. (#70)
+- The `{source|target}_mask_attr_name` field is moved to inside the edge builder definition. (#70)
 
 ## [0.3.0 Anemoi-graphs, minor release](https://github.com/ecmwf/anemoi-graphs/compare/0.2.1...0.3.0) - 2024-09-03